Bring corebos latest changes to AC fork

+++ modules/Users/Save.php
@@ -1,23 +1,19 @@
 <?php
-/*********************************************************************************
- * The contents of this file are subject to the SugarCRM Public License Version 1.1.2
- * ("License"); You may not use this file except in compliance with the 
- * License. You may obtain a copy of the License at http://www.sugarcrm.com/SPL
- * Software distributed under the License is distributed on an  "AS IS"  basis,
- * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for
- * the specific language governing rights and limitations under the License.
- * The Original Code is:  SugarCRM Open Source
- * The Initial Developer of the Original Code is SugarCRM, Inc.
- * Portions created by SugarCRM are Copyright (C) SugarCRM, Inc.;
+/*+**********************************************************************************
+ * The contents of this file are subject to the vtiger CRM Public License Version 1.0
+ * ("License"); You may not use this file except in compliance with the License
+ * The Original Code is:  vtiger CRM Open Source
+ * The Initial Developer of the Original Code is vtiger.
+ * Portions created by vtiger are Copyright (C) vtiger.
  * All Rights Reserved.
- ********************************************************************************/
+ ************************************************************************************/
 require_once('modules/Users/Users.php');
 require_once('include/logging.php');
 require_once('include/utils/UserInfoUtil.php');
-$log =& LoggerManager::getLogger('index');
+$log = LoggerManager::getLogger('index');
 
 global $adb;
-$user_name = vtlib_purify($_REQUEST['userName']);
+$user_name = empty($_REQUEST['userName']) ? '' : vtlib_purify($_REQUEST['userName']);
 if(isset($_REQUEST['status']) && $_REQUEST['status'] != '')
 	$_REQUEST['status']= vtlib_purify ($_REQUEST['status']);
 else
@@ -40,7 +36,7 @@ if(isset($_REQUEST['dup_check']) && $_REQUEST['dup_check'] != '')
 		die;
 	}
 }
-if($_REQUEST['user_role'] != '' && !is_admin($current_user) && $_REQUEST['user_role'] != $current_user->roleid){
+if(!empty($_REQUEST['user_role']) && !is_admin($current_user) && $_REQUEST['user_role'] != $current_user->roleid){
 	$log->fatal("SECURITY:Non-Admin user:". $current_user->id . " attempted to change user role");
 	echo "<link rel='stylesheet' type='text/css' href='themes/$theme/style.css'>";
 	echo "<table border='0' cellpadding='5' cellspacing='0' width='100%' height='450px'><tr><td align='center'>";
@@ -81,26 +77,26 @@ else
 	$focus->mode='';
 }
 
-if($_REQUEST['deleteImage'] == 'true') {
+if(isset($_REQUEST['deleteImage']) && $_REQUEST['deleteImage'] == 'true') {
 	$focus->id = vtlib_purify($_REQUEST['recordid']);
 	$focus->deleteImage();
 	echo "SUCCESS";
 	exit;
 }
 
-if($_REQUEST['changepassword'] == 'true') {
+if(isset($_REQUEST['changepassword']) && $_REQUEST['changepassword'] == 'true') {
 	$focus->retrieve_entity_info($_REQUEST['record'],'Users');
 	$focus->id = vtlib_purify($_REQUEST['record']);
 	if (isset($_REQUEST['new_password'])) {
 		if (!$focus->change_password(vtlib_purify($_REQUEST['old_password']), vtlib_purify($_REQUEST['new_password']))) {
-			header("Location: index.php?action=Error&module=Users&error_string=".urlencode($focus->error_string));
+			header("Location: index.php?action=DetailView&module=Users&record=".$focus->id."&error_string=".urlencode($focus->error_string));
 			exit;
 		}
 	}
 }
 
 //save user Image
-if(! $_REQUEST['changepassword'] == 'true')
+if(empty($_REQUEST['changepassword']) || $_REQUEST['changepassword'] != 'true')
 {
 	if(strtolower($current_user->is_admin) == 'off' && $current_user->id != $focus->id)
 	{
@@ -124,7 +120,7 @@ if(! $_REQUEST['changepassword'] == 'true')
 	else
 		$focus->column_fields['internal_mailer'] = 0;
 	if(isset($_SESSION['internal_mailer']) && $_SESSION['internal_mailer'] != $focus->column_fields['internal_mailer'])
-		$_SESSION['internal_mailer'] = $focus->column_fields['internal_mailer'];
+		coreBOS_Session::set('internal_mailer', $focus->column_fields['internal_mailer']);
 	setObjectValuesFromRequest($focus);
 
 	if(empty($focus->column_fields['roleid']) && !empty($_POST['user_role'])) {
@@ -140,7 +136,7 @@ if(! $_REQUEST['changepassword'] == 'true')
 		$new_pass = md5($new_pass);
 		$uname = $_POST['user_name'];
 		if (!$focus->change_password($_POST['confirm_new_password'], $_POST['new_password'])) {
-			header("Location: index.php?action=Error&module=Users&error_string=".urlencode($focus->error_string));
+			header("Location: index.php?action=DetailView&module=Users&record=".$focus->id."&error_string=".urlencode($focus->error_string));
 			exit;
 		}
 	}
@@ -157,16 +153,16 @@ else $return_module = "Users";
 if(isset($_POST['return_action']) && $_POST['return_action'] != "") $return_action = vtlib_purify($_REQUEST['return_action']);
 else $return_action = "DetailView";
 if(isset($_POST['return_id']) && $_POST['return_id'] != "") $return_id = vtlib_purify($_REQUEST['return_id']);
-if(isset($_REQUEST['activity_mode']))   $activitymode = '&activity_mode='.vtlib_purify($_REQUEST['activity_mode']);
 if(isset($_POST['parenttab'])) $parenttab = getParentTab();
 
 $log->debug("Saved record with id of ".$return_id);
 
-//Asha: Added Check to see if the mode is User Creation and if yes, then sending the email notification to the User with Login details.
-if($_REQUEST['mode'] == 'create') {
+// Check to see if the mode is User Creation and if yes, then sending the email notification to the User with Login details.
+$error_str = '';
+if(isset($_REQUEST['mode']) and $_REQUEST['mode'] == 'create') {
 	global $app_strings, $mod_strings, $default_charset;
 	require_once('modules/Emails/mail.php');
-    $user_emailid = $focus->column_fields['email1'];
+	$user_emailid = $focus->column_fields['email1'];
 	// send email on Create user only if NOTIFY_OWNER_EMAILS is set to true
 
 	$subject = $mod_strings['User Login Details'];
@@ -179,21 +175,19 @@ if($_REQUEST['mode'] == 'create') {
 	$email_body .= "<br>" . $app_strings['MSG_THANKS'] . "<br>" . $current_user->user_name;
 	//$email_body = htmlentities($email_body, ENT_QUOTES, $default_charset);  // not needed anymore, PHPMailer takes care of it
 
+	$HELPDESK_SUPPORT_EMAIL_ID = GlobalVariable::getVariable('HelpDesk_Support_EMail','support@your_support_domain.tld','HelpDesk');
+	$HELPDESK_SUPPORT_NAME = GlobalVariable::getVariable('HelpDesk_Support_Name','your-support name','HelpDesk');
 	$mail_status = send_mail('Users',$user_emailid,$HELPDESK_SUPPORT_NAME,$HELPDESK_SUPPORT_EMAIL_ID,$subject,$email_body);
 	if($mail_status != 1) {
 		$mail_status_str = $user_emailid."=".$mail_status."&&&";
 		$error_str = getMailErrorString($mail_status_str);
 	}
 }
-$location = "Location: index.php?action=".vtlib_purify($return_action)."&module=".vtlib_purify($return_module)."&record=".vtlib_purify($return_id);
-
-if($_REQUEST['modechk'] != 'prefview') {
-	$location .= "&parenttab=".vtlib_purify($parenttab);
-}
+$location = 'Location: index.php?action='.urlencode(vtlib_purify($return_action)).'&module='.urlencode(vtlib_purify($return_module)).'&record='.urlencode(vtlib_purify($return_id));
 
 if ($error_str != '') {
 	$user = $focus->column_fields['user_name'];
-	$location .= "&user=$user&$error_str";
+	$location .= '&user=' . urlencode($user) . '&' . $error_str;
 }
 
 header($location);
